fix: address CLI forwarding review feedback
